refactor(sefariasqlite): extract generation seeding to its own task
Moves generation seeding out of renameCategories into a new
:sefariasqlite:seedGenerations task that runs after appendOtzaria. Same
matcher, same CSV, same junction — but the seed pass now sees Otzaria-stage
books too. Per-book CSV granularity is preserved (no transitive author-level
propagation that would flatten the empty-author bucket or אברבנאל's
ראשונים/אחרונים split).

Shared CSV helpers in RenameCategoriesPostProcess.kt (parsePairs,
downloadCsv, stripTitlePunct, STRIP_TITLE_PUNCT_SQL, FOR_DB_BASE) promoted
from private to internal so the new file reuses them without duplication.

Local result: unlinked books drop from 1559 → 697. The Otzaria-curated
sources are now nearly fully linked (OnYourWayToOtzaria 99 → 0, Dicta
658 → 121, ToratEmet 97 → 9). Remaining unlinked are books legitimately
absent from the CSV.
